+++ sdk/serialization/azure-json/src/main/java/com/azure/json/JsonReader.java
@@ -8,6 +8,8 @@ import com.azure.json.implementation.jackson.core.io.JsonStringEncoder;
 
 import java.io.Closeable;
 import java.io.IOException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Base64;
@@ -567,8 +569,8 @@ public abstract class JsonReader implements Closeable {
      * <ul>
      *     <li>null if the starting token is null or {@link JsonToken#NULL}</li>
      *     <li>true or false if the starting token is {@link JsonToken#BOOLEAN}</li>
-     *     <li>One of int, long, float, or double is the starting token is {@link JsonToken#NUMBER}, the smallest
-     *     containing value will be used if the number is an integer</li>
+     *     <li>One of int, long, {@link BigInteger}, double, or {@link BigDecimal} if the starting token is
+     *     {@link JsonToken#NUMBER}, the smallest containing value will be used</li>
      *     <li>An array of untyped elements if the starting point is {@link JsonToken#START_ARRAY}</li>
      *     <li>A map of String-untyped value if the starting point is {@link JsonToken#START_OBJECT}</li>
      * </ul>