+++ sdk/serialization/azure-json/src/test/java/com/azure/json/contract/JsonReaderContractTests.java
@@ -202,7 +202,7 @@ public abstract class JsonReaderContractTests {
             assertEquals("string", jsonArray[0]);
             assertNull(jsonArray[1]);
             assertEquals(10, jsonArray[2]);
-            assertEquals(10.0F, jsonArray[3]);
+            assertEquals(10.0D, jsonArray[3]);
             assertEquals(true, jsonArray[4]);
         }
     }
@@ -366,7 +366,7 @@ public abstract class JsonReaderContractTests {
             assertEquals("string", jsonArray[0]);
             assertNull(jsonArray[1]);
             assertEquals(10, jsonArray[2]);
-            assertEquals(10.0F, jsonArray[3]);
+            assertEquals(10.0D, jsonArray[3]);
             assertEquals(true, jsonArray[4]);
             assertEquals("innerString", jsonArray[5]);
         }
@@ -413,7 +413,7 @@ public abstract class JsonReaderContractTests {
             assertEquals("string", jsonArray[0]);
             assertNull(jsonArray[1]);
             assertEquals(10, jsonArray[2]);
-            assertEquals(10.0F, jsonArray[3]);
+            assertEquals(10.0D, jsonArray[3]);
             assertEquals(true, jsonArray[4]);
             assertEquals("innerString", jsonArray[5]);
         }
@@ -447,7 +447,7 @@ public abstract class JsonReaderContractTests {
 
     private static Stream<Arguments> readUntypedSimpleSupplier() {
         return Stream.of(Arguments.of("null", 1, null), Arguments.of("true", 1, true), Arguments.of("false", 1, false),
-            Arguments.of("3.14", 1, 3.14F), Arguments.of("NaN", 1, Double.NaN),
+            Arguments.of("3.14", 1, 3.14D), Arguments.of("NaN", 1, Double.NaN),
             Arguments.of("-Infinity", 1, Double.NEGATIVE_INFINITY),
             Arguments.of("Infinity", 1, Double.POSITIVE_INFINITY), Arguments.of("42", 1, 42),
             Arguments.of("420000000000", 1, 420000000000L), Arguments.of("\"hello\"", 1, "hello"));
@@ -710,7 +710,7 @@ public abstract class JsonReaderContractTests {
         ReadValueCallback<JsonReader, List<Object>> reader = read(r -> r.readArray(JsonReader::readUntyped));
         return Stream.of(Arguments.of("null", reader, null), Arguments.of("[]", reader, Collections.emptyList()),
             Arguments.of("[10]", reader, Collections.singletonList(10)),
-            Arguments.of("[true,10,10.0,\"hello\"]", reader, Arrays.asList(true, 10, 10.0F, "hello")));
+            Arguments.of("[true,10,10.0,\"hello\"]", reader, Arrays.asList(true, 10, 10.0D, "hello")));
     }
 
     @Test
@@ -740,7 +740,7 @@ public abstract class JsonReaderContractTests {
         Map<String, Object> map = new LinkedHashMap<>();
         map.put("boolean", true);
         map.put("int", 42);
-        map.put("decimal", 42.0F);
+        map.put("decimal", 42.0D);
         map.put("string", "hello");
 
         return Stream.of(Arguments.of("null", reader, null), Arguments.of("{}", reader, Collections.emptyMap()),
@@ -755,9 +755,9 @@ public abstract class JsonReaderContractTests {
     }
 
     private static Stream<Arguments> readUntypedExponentNumbersSupplier() {
-        return Stream.of(Arguments.of("1e-1", 0.1F), Arguments.of("1E-1", 0.1F), Arguments.of("1e+1", 10F),
-            Arguments.of("1E+1", 10F), Arguments.of("1e-01", 0.1F), Arguments.of("1E-01", 0.1F),
-            Arguments.of("1e+01", 10F), Arguments.of("1E+01", 10F), Arguments.of("1e0", 1F), Arguments.of("1E0", 1F),
+        return Stream.of(Arguments.of("1e-1", 0.1D), Arguments.of("1E-1", 0.1D), Arguments.of("1e+1", 10D),
+            Arguments.of("1E+1", 10D), Arguments.of("1e-01", 0.1D), Arguments.of("1E-01", 0.1D),
+            Arguments.of("1e+01", 10D), Arguments.of("1E+01", 10D), Arguments.of("1e0", 1D), Arguments.of("1E0", 1D),
 
             Arguments.of("Infinity", Double.POSITIVE_INFINITY), Arguments.of("+Infinity", Double.POSITIVE_INFINITY),
             Arguments.of("-Infinity", Double.NEGATIVE_INFINITY), Arguments.of("NaN", Double.NaN));
@@ -790,9 +790,8 @@ public abstract class JsonReaderContractTests {
             return test;
         }
 
-        public TestData setTest(String test) {
+        public void setTest(String test) {
             this.test = test;
-            return this;
         }
 
         @Override