+++ sdk/serialization/azure-json/CHANGELOG.md
@@ -6,14 +6,16 @@
 
 ### Breaking Changes
 
-- `JsonReader.readUntyped()` now returns the smallest possible `Number` when parsing decimals. Previously, `double` was
-  always used. This change is to align with the behavior `JsonNumber`.
+- `JsonNumber` previously would use `float` when the floating point number was small enough to fit in `float` but it
+  now aligns behavior with `JsonReader.readUntyped()` where `double` will be the smallest floating point type used.
+  This aligns with floating point number behavior in Java where `double` is the default if no type is specified.
 - Support for special numeric `INF`, `-INF`, and `+INF` values have been removed to align with behaviors of `Float`
   and `Double` in Java where only the `Infinity` variants are supported.
 
 ### Bugs Fixed
 
-- `JsonReader.readUntyped()` had incomplete support for untyped numerics where exponents were not being parsed correctly.
+- `JsonReader.readUntyped()` had incomplete support for untyped numerics. Numerics too large for `double` and `long` are
+  now supported and a bug where exponents were not being parsed correctly is fixed.
 
 ### Other Changes